MOD LoadBusiness om meer te krijgen

+++ PancakeClicker/PancakeLibrary/Business.cs
@@ -8,7 +8,7 @@ namespace PancakeLibrary
 {
     public class Business
     {
-        public Business(int id, string name, string description, int initialPrice, decimal initialMoneyPerSecond)
+        public Business(int id, string name, string description, decimal initialPrice, decimal initialMoneyPerSecond)
         {
             Id = id;
             Name = name;
@@ -20,7 +20,7 @@ namespace PancakeLibrary
         public int Id { get; set; }
         public string Name { get; set; }
         public string Description { get; set; }
-        public int InitialPrice { get; set; }
+        public decimal InitialPrice { get; set; }
         public decimal InitialMoneyPerSecond { get; set; }
         public uint Amount { get; set; }
     }
